Make a few array entries only as big as they absolutely need to be (trivial).
This decreases the size of the superiotool binary from ca. 1.1 MB to 600 KB.

Acked-by: Uwe Hermann <uwe@hermann-uwe.de>



git-svn-id: svn://svn.coreboot.org/coreboot/trunk@3194 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1

+++ util/superiotool/superiotool.c
@@ -80,7 +80,7 @@ static void dump_regs(const struct superio_registers reg_table[],
 		      int i, int j, uint16_t port)
 {
 	int k;
-	const int *idx;
+	const int16_t *idx;
 
 	if (reg_table[i].ldn[j].ldn != NOLDN) {
 		printf("LDN 0x%02x", reg_table[i].ldn[j].ldn);
